# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaText
  description: | 
    This is the yaml documentation for the WasaText Api
  version: 1.0.0
paths:
  /session:
    post:
      tags:
        - login
      summary: Logs in the user
      description: >
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: Maria
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    example: "1"
...
  /conversations/delete/{conversation_id}:
    delete:
      tags:
        - conversations
      summary: Delete a conversation
      operationId: deleteConversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/conversation_id'
      responses:
        '204':
          description: Conversation deleted successfully
        '404':
          description: Conversation not found
        '403':
          description: Forbidden
...
  /conversations/create-group:
    post:
      tags:
        - groups
      summary: Create a new group
      description: Creates a new group with the specified members.
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        description: Group details to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation_id:
                    type: string
                    example: "1"
        '400':
          description: Invalid request
...
  /conversations/group/add/{group_id}:
    post:
      parameters:
      - $ref: '#/components/parameters/group_id'
      tags:
        - groups
      summary: Add a user to a group
      operationId: addToGroup
      security:
        - bearerAuth: []
      requestBody:
        description: User ID to add
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "hanni"
      responses:
        '200':
          description: User added to group
  /conversations/group/leave/{group_id}:
    delete:
      tags:
        - groups
      summary: Leave a group
      operationId: leaveGroup
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/group_id'
      responses:
        '200':
          description: User left the group successfully
        '404':
          description: Group not found
...
  /user/get-photo/{user_id}:
    get:
      tags:
        - users
      summary: Get the photo of a user
      description: >
        Retrieves the photo of a user by their ID. If the user does not have a photo, a default photo is returned.
      operationId: getUserPhoto
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose photo is to be retrieved.
          schema:
            type: string
      responses:
        '200':
          description: Photo retrieved successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: User not found or photo not set
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
components:
  parameters:
...
    group_id:
      schema:
        type: string
      name: group_id
      in: path
      required: true
      description: The ID of the group
      allowEmptyValue: false
    conversation_id:
      schema:
        type: string
      name: conversation_id
      in: path
      required: true
      description: The conversation ID
      allowEmptyValue: false
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    NewGroup:
      type: object
      properties:
        name:
          type: string
          example: "Group Example"
        members:
          type: array
          items:
            type: string
            example: "hanni"
      required:
        - name
        - members
...